feat: add use case to get installed apps
This commit introduces the `GetInstalledAppsUseCase` class.

The `execute` method in this class retrieves a list of installed applications from the device. It filters out system applications and maps the remaining applications to `AppInfo` objects, which include the app's name, package name, and icon. The resulting list is then sorted alphabetically by app name.